Crux.Structs behaviour (crux_structs v0.3.0-dev)

Provides a unified function to create one or a list of structs, invoking their create/1 function if available.

Functions

create(data, target)
(since 0.1.0)

Specs

create(data :: map(), target :: module()) :: struct()

create(data :: list(), target :: module()) :: [struct()]

Creates a struct or a list of structs invoking their create/1 function if available.

Examples

# A single member
iex> %{
...>   "nick" => "nick",
...>   "user" => %{"username" => "space", "discriminator" => "0001", "id" => "218348062828003328", "avatar" => "646a356e237350bf8b8dfde15667dfc4"},
...>   "roles" => ["251158405832638465", "373405430589816834"],
...>   "mute" => false,
...>   "deaf" => false,
...>   "joined_at" => "2016-11-02T00:51:21.342000+00:00"
...> }
...> |> Crux.Structs.create(Crux.Structs.Member)
%Crux.Structs.Member{
  nick: "nick",
  user: 218348062828003328,
  roles: MapSet.new([251158405832638465, 373405430589816834]),
  mute: false,
  deaf: false,
  joined_at: "2016-11-02T00:51:21.342000+00:00",
  guild_id: nil
}

# A single user
iex> %{"username" => "space", "discriminator" => "0001", "id" => "218348062828003328", "avatar" => "46a356e237350bf8b8dfde15667dfc4"}
...> |> Crux.Structs.create(Crux.Structs.User)
%Crux.Structs.User{username: "space", discriminator: "0001", id: 218348062828003328, avatar: "46a356e237350bf8b8dfde15667dfc4", bot: false, system: false}

# Multiple users
iex> [
...>   %{"username" => "space", "discriminator" => "0001", "id" => "218348062828003328", "avatar" => "46a356e237350bf8b8dfde15667dfc4"},
...>   %{"username" => "Drahcirius", "discriminator" => "1336", "id" => "130175406673231873", "avatar" => "c896aebec82c90f590b08cfebcdc4e3b"}
...> ]
...> |> Crux.Structs.create(Crux.Structs.User)
[
  %Crux.Structs.User{username: "space", discriminator: "0001", id: 218348062828003328, avatar: "46a356e237350bf8b8dfde15667dfc4", bot: false, system: false},
  %Crux.Structs.User{username: "Drahcirius", discriminator: "1336", id: 130175406673231873, avatar: "c896aebec82c90f590b08cfebcdc4e3b", bot: false, system: false}
]

# Does not alter already structs
iex> Crux.Structs.create(
...>   %Crux.Structs.User{username: "space", discriminator: "0001", id: 218348062828003328, avatar: "46a356e237350bf8b8dfde15667dfc4", bot: false, system: false},
...>   Crux.Structs.User
...> )
%Crux.Structs.User{username: "space", discriminator: "0001", id: 218348062828003328, avatar: "46a356e237350bf8b8dfde15667dfc4", bot: false, system: false}

# Fallback
iex> Crux.Structs.create(nil, nil)
nil

resolve_id(data, target)
(since 0.2.1)

Specs

resolve_id(nil, target :: module()) :: nil

resolve_id(Crux.Structs.Snowflake.t(), target :: module()) ::
  Crux.Structs.Snowflake.t()

resolve_id(map(), target :: module()) :: Crux.Structs.Snowflake.t() | nil

Resolves the id of a struct invoking their resolve_id/1 function if available.

# Struct of the concrete type
iex> %Crux.Structs.Webhook{id: 618733351624507394}
...> |> Crux.Structs.resolve_id(Crux.Structs.Webhook)
618733351624507394

# Already snowflake
iex> 222089067028807682
...> |> Crux.Structs.resolve_id(Crux.Structs.Role)
222089067028807682

# Snowflake string
iex> "222079895583457280"
...> |> Crux.Structs.resolve_id(Crux.Structs.Channel)
222079895583457280

# nil
iex> nil
...> |> Crux.Structs.resolve_id(Crux.Structs.Guild)
nil

# Inexact type that is a resolvable
iex> %Crux.Structs.Member{user: 218348062828003328}
...> |> Crux.Structs.resolve_id(Crux.Structs.User)
218348062828003328

# Incorrect type
iex> %Crux.Structs.Role{id: 222079439876390922}
...> |> Crux.Structs.resolve_id(Crux.Structs.Emoji)
nil

Callbacks

create(data)
(optional)

Specs

create(data :: map()) :: struct()

Can be implemented by structs to transform the inital data.

resolve_id(data)
(optional)

Specs

resolve_id(data :: map() | Crux.Structs.Snowflake.t()) ::
  Crux.Structs.Snowflake.t() | nil

Can be implemented by structs to provide a mechanism to resolve their id.