Crux.Structs.AuditLogEntry (crux_structs v0.3.0-dev)

Represents a Discord Audit Log Entry Object.

Types

event_name()
(since 0.1.6)

Specs

event_name() ::
  :guild_update
  | :channel_create
  | :channel_update
  | :channel_delete
  | :channel_overwrite_create
  | :channel_overwrite_update
  | :channel_overwrite_delete
  | :member_kick
  | :member_prune
  | :member_ban_add
  | :member_ban_remove
  | :member_update
  | :member_role_update
  | :member_move
  | :member_disconnect
  | :bot_add
  | :role_create
  | :role_update
  | :role_delete
  | :invite_create
  | :invite_update
  | :invite_delete
  | :webhook_create
  | :webhook_update
  | :webhook_delete
  | :emoji_create
  | :emoji_update
  | :emoji_delete
  | :message_delete
  | :message_bulk_delete
  | :message_pin
  | :message_unpin
  | :integration_create
  | :integration_update
  | :integration_delete

Union type of audit log event name atoms.

options()
(since 0.3.0)

Specs

options() :: %{
  optional(:delete_member_days) => non_neg_integer(),
  optional(:members_removed) => non_neg_integer(),
  optional(:channel_id) => Crux.Structs.Snowflake.t(),
  optional(:message_id) => Crux.Structs.Snowflake.t(),
  optional(:count) => non_neg_integer(),
  optional(:id) => Crux.Structs.Snowflake.t(),
  optional(:type) => non_neg_integer(),
  optional(:role_name) => String.t()
}

Additional info for certain action types.

For more information see the Discord Developer Documentation.

t()
(since 0.1.6)

Specs

t() :: %Crux.Structs.AuditLogEntry{
  action_type: non_neg_integer(),
  changes: %{required(String.t()) => Crux.Structs.AuditLogChange.t()},
  id: Crux.Structs.Snowflake.t(),
  options: options() | nil,
  reason: String.t() | nil,
  target_id: Crux.Structs.Snowflake.t(),
  user_id: Crux.Structs.Snowflake.t()
}

Functions

create(data)
(since 0.1.6)

Specs

create(data :: map()) :: t()

Creates a Crux.Structs.AuditLogEntry.t/0 struct from raw data.

Automatically invoked by Crux.Structs.create/2.

event_name(action_type)
(since 0.1.6)

Specs

event_name(action_type :: non_neg_integer()) :: atom()

Gets the event name from the action type id

events()
(since 0.1.6)

Specs

events() :: %{required(event_name()) => non_neg_integer()}

Returns a map of all audit log event names with their id