Crux.Structs.Channel (crux_structs v0.3.0-dev)

Represents a Discord Channel Object.

Types

id_resolvable()
(since 0.2.1)

Specs

id_resolvable() ::
  Crux.Structs.Message.t() | t() | Crux.Structs.Snowflake.t() | String.t()

All available types that can be resolved into a channel id.

position_resolvable()
(since 0.2.1)

Specs

position_resolvable() ::
  t()
  | %{channel: id_resolvable(), position: integer()}
  | {id_resolvable(), integer()}
  | %{id: id_resolvable(), position: integer()}

All available types that can be resolved into a channel position.

t()
(since 0.1.0)

Specs

t() :: %Crux.Structs.Channel{
  application_id: Crux.Structs.Snowflake.t(),
  bitrate: integer(),
  default_auto_archive_duration: 60 | 1440 | 4320 | 10080,
  guild_id: Crux.Structs.Snowflake.t(),
  icon: String.t(),
  id: Crux.Structs.Snowflake.t(),
  last_message_id: Crux.Structs.Snowflake.t(),
  last_pin_timestamp: String.t(),
  member: thread_member(),
  member_count: 1..50,
  message_count: 1..50,
  name: String.t(),
  nsfw: boolean(),
  owner_id: Crux.Structs.Snowflake.t(),
  parent_id: Crux.Structs.Snowflake.t(),
  permission_overwrites: %{
    optional(Crux.Structs.Snowflake.t()) => Crux.Structs.Overwrite.t()
  },
  position: integer(),
  rate_limit_per_user: integer(),
  recipients: MapSet.t(Crux.Structs.Snowflake.t()),
  rtc_region: String.t(),
  thread_metadata: thread_metadata(),
  topic: String.t(),
  type: type(),
  user_limit: non_neg_integer(),
  video_quality_mode: 1 | 2
}

thread_member()
(since 0.3.0)

Specs

thread_member() :: %{
  id: Crux.Structs.Snowflake.t(),
  user_id: Crux.Structs.Snowflake.t(),
  join_timestamp: String.t(),
  flags: integer()
}

A thread member.

Notes:

:id -> Id of the thread, not member / user

:user_id -> Acutal id of the member / user

For more information see the Discord Developer Documentation.

thread_metadata()
(since 0.3.0)

Specs

thread_metadata() :: %{
  optional(:invitable) => boolean(),
  :archived => boolean(),
  :auto_archive_duration => 60 | 1440 | 4320 | 10080,
  :archive_timestamp => String.t(),
  :locked => boolean()
}

Additional metadata for threads.

For more information see the Discord Developer Documentation.

type()
(since 0.2.3)

Specs

type() :: non_neg_integer()

The type of a channel.

Type	ID	Description

GUILD_TEXT	0	A text channel within a guild.

DM	1	A direct text channel between two users.

GUILD_VOICE	2	A voice channel withing a guild.

GROUP_DM	3	A direct channel between multiple users. Bots do not have access to those.

GUILD_CATEGORY	4	An organizational category.

GUILD_NEWS	5	A text channel users can follow and crosspost messages to.

GUILD_STORE	6	A channel in which game developers can sell their game. Bots can not interact with those.

GUILD_NEWS_THREAD	10	A temporary sub-channel within a GUILD_NEWS channel

GUILD_PUBLIC_THREAD	11	A temporay sub-channel within a GUILD_TEXT channel

GUILD_PRIVATE_THREAD	12	A temporary sub-channel within a GUILD_TEXT channel only viewable by invited and members with MANAGE_THREADS permission

GUILD_STAGE_VOICE	13	A voice channel for hosting events with an audience

For more information see the Discord Developer Documentation.

Functions

create(data)

Specs

create(data :: map()) :: t()

Creates a Crux.Structs.Channel.t/0 struct from raw data.

Automatically invoked by Crux.Structs.create/2

resolve_id(resolvable)
(since 0.2.3)

Specs

resolve_id(id_resolvable()) :: Crux.Structs.Snowflake.t() | nil

Resolves the id of a Crux.Structs.Channel.t/0.

Automatically invoked by Crux.Structs.resolve_id/2.

iex> %Crux.Structs.Message{channel_id: 222079895583457280}
...> |> Crux.Structs.Channel.resolve_id()
222079895583457280

iex> %Crux.Structs.Channel{id: 222079895583457280}
...> |> Crux.Structs.Channel.resolve_id()
222079895583457280

iex> 222079895583457280
...> |> Crux.Structs.Channel.resolve_id()
222079895583457280

iex> "222079895583457280"
...> |> Crux.Structs.Channel.resolve_id()
222079895583457280

resolve_position(resolvable)
(since 0.2.1)

Specs

resolve_position(position_resolvable()) ::
  %{id: Crux.Structs.Snowflake.t(), position: integer()} | nil

Resolves a position_resolvable/0 into a channel position.

Examples

iex> %Crux.Structs.Channel{id: 222079895583457280, position: 5}
...> |> Crux.Structs.Channel.resolve_position()
%{id: 222079895583457280, position: 5}

iex> {%Crux.Structs.Channel{id: 222079895583457280}, 5}
...> |> Crux.Structs.Channel.resolve_position()
%{id: 222079895583457280, position: 5}

iex> {222079895583457280, 5}
...> |> Crux.Structs.Channel.resolve_position()
%{id: 222079895583457280, position: 5}

iex> %{id: 222079895583457280, position: 5}
...> |> Crux.Structs.Channel.resolve_position()
%{id: 222079895583457280, position: 5}

iex> %{channel: 222079895583457280, position: 5}
...> |> Crux.Structs.Channel.resolve_position()
%{id: 222079895583457280, position: 5}

iex> {nil, 5}
...> |> Crux.Structs.Channel.resolve_position()
nil

to_mention(channel)
(since 0.1.1)

Specs

to_mention(user :: t()) :: String.t()

Converts a Crux.Structs.Channel.t/0 into its discord mention format.

Example

iex> %Crux.Structs.Channel{id: 316880197314019329}
...> |> Crux.Structs.Channel.to_mention()
"<#316880197314019329>"