Crux.Structs.Guild.SystemChannelFlags (crux_structs v0.3.0-dev)

Custom non discord API struct helping with the usage of system channel flags.

Discord is using flags in an "inverted" way here to allow all new types to be enabled by default. Therefore all set flags are actually disabled (as their names suggest).

Types

name()
(since 0.2.3)

Specs

name() :: :suppress_join_notifications | :suppress_premium_subscriptions

Union type of all valid system channel flags.

raw()

Specs

raw() :: non_neg_integer() | String.t()

Raw bitfield that can be used as a resolvable/0.

resolvable()

Specs

resolvable() :: t() | raw() | name() | [resolvable()]

All valid types that can be directly resolved into a bitfield.

t()

Specs

t() :: non_neg_integer()

Functions

add(base, to_add)

Specs

add(base :: resolvable(), to_add :: resolvable()) :: t()

Add all set bits of to_add to base.

all()

Specs

all() :: t()

Get an integer representing all available bits set.

flags()

Specs

flags() :: %{required(name()) => t()}

Get a map of name/0s and their corresponding bit values.

has(have, want)

Specs

has(have :: resolvable(), want :: resolvable()) :: boolean()

Check whether the resolvable/0 you have has everything set you want.

missing(have, want)

Specs

missing(resolvable(), resolvable()) :: t()

Return a t/0 of all bits you want but not have.

names()

Specs

names() :: [name()]

Get a list of all available name/0s.

remove(base, to_remove)

Specs

remove(base :: resolvable(), to_remove :: resolvable()) :: t()

Remove all set bits of to_remove from base.

resolve(resolvable)

Specs

resolve(resolvable()) :: t()

Resolve a resolvable/0 into a bitfield.

to_list(resolvable)

Specs

to_list(resolvable()) :: [name()]

Serialize a resolvable/0 into a list of set bit flag names.

to_map(resolvable)

Specs

to_map(resolvable()) :: %{required(name()) => boolean()}

Serialize a resolvable/0 into a map representing bit flag names to whether they are set.