Crux.Structs.Snowflake (crux_structs v0.3.0-dev)

Custom non discord api struct to help with working with Discord's snowflakes.

For more information see Discord Docs.

Types

resolvable()
(since 0.2.1)

Specs

resolvable() :: String.t() | t()

All valid types that can be resolved into a t/0.

t()
(since 0.2.1)

Specs

t() :: 0..18_446_744_073_709_551_615

A discord snowflake, an unsigned 64 bit integer.

Functions

construct(parts)
(since 0.2.1)

Specs

construct(Crux.Structs.Snowflake.Parts.t() | Keyword.t()) :: t()

Constructs a t/0 from its Crux.Structs.Snowflake.Parts.t/0 or a keyword of its fields.

iex> %Crux.Structs.Snowflake.Parts{increment: 0, process_id: 0, timestamp: 1472128634889, worker_id: 1}
...> |> Crux.Structs.Snowflake.construct()
218348062828003328

iex> Crux.Structs.Snowflake.construct(increment: 1, timestamp: 1451106635493)
130175406673231873

iex> Crux.Structs.Snowflake.construct(timestamp: Crux.Structs.Snowflake.discord_epoch())
0

deconstruct(snowflake)
(since 0.2.1)

Specs

deconstruct(t()) :: Crux.Structs.Snowflake.Parts.t()

Deconstructs a t/0 to its Crux.Structs.Snowflake.Parts.t/0.

iex> Crux.Structs.Snowflake.deconstruct(218348062828003328)
%Crux.Structs.Snowflake.Parts{
  increment: 0,
  process_id: 0,
  timestamp: 1472128634889,
  worker_id: 1
}

discord_epoch()
(since 0.2.1)

Specs

discord_epoch() :: non_neg_integer()

The discord epoch, the first second of 2015 or 1420070400000.

iex> Crux.Structs.Snowflake.discord_epoch()
1_420_070_400_000

from_integer(snowflake)
(since 0.2.1)

Specs

from_integer(t()) :: Crux.Structs.Snowflake.Parts.t()

Deconstructs a t/0 to its Crux.Structs.Snowflake.Parts.t/0.

is_snowflake(snowflake)
(macro) (since 0.2.1)

Returns true if term is a t/0; otherwise returns false..

parse(snowflake)
(since 0.2.1)

Specs

parse(t()) :: t()

parse(String.t()) :: t() | :error

Converts a String.t/0 to a t/0 while allowing t/0 to pass through.

Returns :error if the provided string is not a t/0.

  iex> Crux.Structs.Snowflake.parse("invalid")
  :error

  iex> Crux.Structs.Snowflake.parse(218348062828003328)
  218348062828003328

  # Fallbacks
  iex> Crux.Structs.Snowflake.parse("218348062828003328")
  218348062828003328

to_integer(t)
(since 0.2.1)

Specs

to_integer(Crux.Structs.Snowflake.Parts.t()) :: t()

Constructs a t/0 from its Crux.Structs.Snowflake.Parts.t/0.

to_snowflake(snowflake)
(since 0.2.1)

Specs

to_snowflake(t()) :: t()

to_snowflake(String.t()) :: t() | no_return()

to_snowflake(nil) :: nil

Converts a String.t/0 to a t/0 while allowing t/0 and nil to pass through.

Raises an ArgumentError if the provided string is not an integer.

  iex> Crux.Structs.Snowflake.to_snowflake(218348062828003328)
  218348062828003328

  # Fallbacks
  iex> Crux.Structs.Snowflake.to_snowflake("218348062828003328")
  218348062828003328

  iex> Crux.Structs.Snowflake.to_snowflake(nil)
  nil