Crux.Structs.Webhook (crux_structs v0.3.0-dev)

Represents a Discord Webhook Object

Differences opposed to the Discord API Object:

:user is just the user id

Types

id_resolvable()
(since 0.2.1)

Specs

id_resolvable() :: t() | Crux.Structs.Snowflake.t() | String.t()

All available types that can be resolved into a webhook id.

t()

Specs

t() :: %Crux.Structs.Webhook{
  application_id: term(),
  avatar: String.t() | nil,
  channel_id: Crux.Structs.Snowflake.t(),
  guild_id: Crux.Structs.Snowflake.t() | nil,
  id: Crux.Structs.Snowflake.t(),
  name: String.t() | nil,
  token: String.t() | nil,
  type: 1..2,
  user: Crux.Structs.Snowflake.t() | nil
}

Functions

create(data)
(since 0.1.6)

Specs

create(data :: map()) :: t()

Creates a Crux.Structs.Webhook.t/0 struct from raw data.

Automatically invoked by Crux.Structs.create/2.

to_mention(webhook)
(since 0.1.6)

Specs

to_mention(webhook :: t()) :: String.t()

Converts a Crux.Structs.Webhook.t/0 into its discord mention format.

Although the discord client does not autocomplete it for you, mentioning one still works.

iex> %Crux.Structs.Webhook{id: 218348062828003328}
...> |> Crux.Structs.Webhook.to_mention()
"<@218348062828003328>"